make schema support nested composition
